Alternative approach where we put this on the options object
